Add files via upload
These are for students on MITx 6.00.2x, and are useless to anyone else. You need the version to match your local Python version (_38 for 3.8, etc etc), and you need to load it in the same directory as the ones already supplied (currently _35 and _36).
Then, in the code provided for the problem sets, make sure your version, and only your version, is uncommented. All other should begin with #.